// src/LupiraCalWeb.Client/src/data/api/models/calendarKind.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * LupiraCalApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type CalendarKind = typeof CalendarKind[keyof typeof CalendarKind];


export const CalendarKind = {
  Personal: 'Personal',
  Group: 'Group',
  Birthdays: 'Birthdays',
  Availability: 'Availability',
  Inbox: 'Inbox',
  LlmPrompts: 'LlmPrompts',
  UserCheckIn: 'UserCheckIn',
  DevOps: 'DevOps',
  FoodPlan: 'FoodPlan',
  Generic: 'Generic',
} as const;