// src/LupiraCalWeb.Client/src/data/api/models/containerDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * LupiraCalApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Access } from './access';
import type { CalendarClass } from './calendarClass';
import type { CalendarKind } from './calendarKind';

export interface ContainerDto {
  id: string;
  type: string;
  slug: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  defaultTimezone?: string | null;
  class?: null | CalendarClass;
  kind?: null | CalendarKind;
  access: Access;
}